+++ LViewPlus64/LeagueMemoryReader.h
@@ -39,7 +39,7 @@ private:
 							   
 	// Memory related		   
 	DWORD_PTR                   moduleBaseAddr    = 0;
-	DWORD                       moduleSize        = 0;
+	DWORD64                       moduleSize        = 0;
 	BOOL                        is64Bit           = FALSE;
 
 private:
