+++ LViewPlus64/MemoryLoadable.h
@@ -4,5 +4,5 @@
 /// Interface to be implemented by game objects that are read from memory
 class MemoryLoadable {
 
-	virtual void LoadFromMem(DWORD base, HANDLE hProcess, bool deepLoad = true) = 0;
+	virtual void LoadFromMem(DWORD64 base, HANDLE hProcess, bool deepLoad = true) = 0;
 };
